requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Authlib==1.2.1
PyMySQL==1.1.0
python-dotenv==1.0.0
pytest==7.4.3